Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2
httpx==0.26.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[recommended]==1.7.4
